--- app/controllers/chatops/slack_slash_command_controller.rb
@@ -1,7 +1,0 @@
-module Chatops
-  class SlackSlashCommandController < ::ApplicationChatopsController
-    def handle
-      render plain: "pong"
-    end
-  end
-end
