+++ app/lib/constraints/matches_command.rb
@@ -1,0 +1,11 @@
+module Constraints
+  class MatchesCommand
+    def initialize(command:)
+      @command = command
+    end
+
+    def matches?(request)
+      @command == request.params["text"]
+    end
+  end
+end
